+++ app/src/main/java/co/develhope/meteoapp/WeekItems.kt
@@ -1,6 +1,7 @@
 package co.develhope.meteoapp
 
 sealed class WeekItems(){
+    //today in questo caso è ridondante anche alla luce dei commenti negl ialtri files
     data class Today(
         val dayOfWeek: String,
         val date: String,
@@ -19,8 +20,10 @@ sealed class WeekItems(){
         val precipitation: Int,
         val windSpeed: Int,
     ): WeekItems()
-
+//il nome delle classi hanno la lettera iniziale maiuscola
     object title : WeekItems()
+    //il nome delle classi hanno la lettera iniziale maiuscola
+
     object subtitle : WeekItems()
 }
